GradientBox component. Add it to Trade page.
